refactor(spec): split temporal.ts into parse, column, and facade
Separate value-level temporal engines from column inference and keep a thin
authoring-helper facade so domain edits do not require reading the full
module. Public package exports and ./temporal.js re-exports are unchanged.
